+++ src/main/java/net/minecraft/entity/DataWatcher.java
@@ -40,12 +40,12 @@ public class DataWatcher {
             throw new IllegalArgumentException("Unknown data type: " + object.getClass());
         } else if (id > 31) {
             throw new IllegalArgumentException("Data value id is too big with " + id + "! (Max is " + 31 + ")");
-        } else if (this.watchedObjects.containsKey(Integer.valueOf(id))) {
+        } else if (this.watchedObjects.containsKey(id)) {
             throw new IllegalArgumentException("Duplicate id value for " + id + "!");
         } else {
             DataWatcher.WatchableObject datawatcher$watchableobject = new DataWatcher.WatchableObject(integer.intValue(), id, object);
             this.lock.writeLock().lock();
-            this.watchedObjects.put(Integer.valueOf(id), datawatcher$watchableobject);
+            this.watchedObjects.put(id, datawatcher$watchableobject);
             this.lock.writeLock().unlock();
             this.isBlank = false;
         }
@@ -54,7 +54,7 @@ public class DataWatcher {
     public void addObjectByDataType(int id, int type) {
         DataWatcher.WatchableObject datawatcher$watchableobject = new DataWatcher.WatchableObject(type, id, null);
         this.lock.writeLock().lock();
-        this.watchedObjects.put(Integer.valueOf(id), datawatcher$watchableobject);
+        this.watchedObjects.put(id, datawatcher$watchableobject);
         this.lock.writeLock().unlock();
         this.isBlank = false;
     }
@@ -88,11 +88,11 @@ public class DataWatcher {
         DataWatcher.WatchableObject datawatcher$watchableobject;
 
         try {
-            datawatcher$watchableobject = this.watchedObjects.get(Integer.valueOf(id));
+            datawatcher$watchableobject = this.watchedObjects.get(id);
         } catch (Throwable throwable) {
             CrashReport crashreport = CrashReport.makeCrashReport(throwable, "Getting synched entity data");
             CrashReportCategory crashreportcategory = crashreport.makeCategory("Synched entity data");
-            crashreportcategory.addCrashSection("Data ID", Integer.valueOf(id));
+            crashreportcategory.addCrashSection("Data ID", id);
             throw new ReportedException(crashreport);
         }
 
@@ -245,19 +245,19 @@ public class DataWatcher {
 
             switch (j) {
                 case 0:
-                    datawatcher$watchableobject = new DataWatcher.WatchableObject(j, k, Byte.valueOf(buffer.readByte()));
+                    datawatcher$watchableobject = new DataWatcher.WatchableObject(j, k, buffer.readByte());
                     break;
 
                 case 1:
-                    datawatcher$watchableobject = new DataWatcher.WatchableObject(j, k, Short.valueOf(buffer.readShort()));
+                    datawatcher$watchableobject = new DataWatcher.WatchableObject(j, k, buffer.readShort());
                     break;
 
                 case 2:
-                    datawatcher$watchableobject = new DataWatcher.WatchableObject(j, k, Integer.valueOf(buffer.readInt()));
+                    datawatcher$watchableobject = new DataWatcher.WatchableObject(j, k, buffer.readInt());
                     break;
 
                 case 3:
-                    datawatcher$watchableobject = new DataWatcher.WatchableObject(j, k, Float.valueOf(buffer.readFloat()));
+                    datawatcher$watchableobject = new DataWatcher.WatchableObject(j, k, buffer.readFloat());
                     break;
 
                 case 4:
@@ -292,7 +292,7 @@ public class DataWatcher {
         this.lock.writeLock().lock();
 
         for (DataWatcher.WatchableObject datawatcher$watchableobject : p_75687_1_) {
-            DataWatcher.WatchableObject datawatcher$watchableobject1 = this.watchedObjects.get(Integer.valueOf(datawatcher$watchableobject.getDataValueId()));
+            DataWatcher.WatchableObject datawatcher$watchableobject1 = this.watchedObjects.get(datawatcher$watchableobject.getDataValueId());
 
             if (datawatcher$watchableobject1 != null) {
                 datawatcher$watchableobject1.setObject(datawatcher$watchableobject.getObject());
@@ -313,14 +313,14 @@ public class DataWatcher {
     }
 
     static {
-        dataTypes.put(Byte.class, Integer.valueOf(0));
-        dataTypes.put(Short.class, Integer.valueOf(1));
-        dataTypes.put(Integer.class, Integer.valueOf(2));
-        dataTypes.put(Float.class, Integer.valueOf(3));
-        dataTypes.put(String.class, Integer.valueOf(4));
-        dataTypes.put(ItemStack.class, Integer.valueOf(5));
-        dataTypes.put(BlockPos.class, Integer.valueOf(6));
-        dataTypes.put(Rotations.class, Integer.valueOf(7));
+        dataTypes.put(Byte.class, 0);
+        dataTypes.put(Short.class, 1);
+        dataTypes.put(Integer.class, 2);
+        dataTypes.put(Float.class, 3);
+        dataTypes.put(String.class, 4);
+        dataTypes.put(ItemStack.class, 5);
+        dataTypes.put(BlockPos.class, 6);
+        dataTypes.put(Rotations.class, 7);
     }
 
     public static class WatchableObject {
