+++ src/main/java/net/minecraft/entity/item/EntityArmorStand.java
@@ -68,7 +68,7 @@ public class EntityArmorStand extends EntityLivingBase {
 
     protected void entityInit() {
         super.entityInit();
-        this.dataWatcher.addObject(10, Byte.valueOf((byte) 0));
+        this.dataWatcher.addObject(10, (byte) 0);
         this.dataWatcher.addObject(11, DEFAULT_HEAD_ROTATION);
         this.dataWatcher.addObject(12, DEFAULT_BODY_ROTATION);
         this.dataWatcher.addObject(13, DEFAULT_LEFTARM_ROTATION);
@@ -596,7 +596,7 @@ public class EntityArmorStand extends EntityLivingBase {
             b0 = (byte) (b0 & -2);
         }
 
-        this.dataWatcher.updateObject(10, Byte.valueOf(b0));
+        this.dataWatcher.updateObject(10, b0);
     }
 
     public boolean isSmall() {
@@ -612,7 +612,7 @@ public class EntityArmorStand extends EntityLivingBase {
             b0 = (byte) (b0 & -3);
         }
 
-        this.dataWatcher.updateObject(10, Byte.valueOf(b0));
+        this.dataWatcher.updateObject(10, b0);
     }
 
     public boolean hasNoGravity() {
@@ -628,7 +628,7 @@ public class EntityArmorStand extends EntityLivingBase {
             b0 = (byte) (b0 & -5);
         }
 
-        this.dataWatcher.updateObject(10, Byte.valueOf(b0));
+        this.dataWatcher.updateObject(10, b0);
     }
 
     public boolean getShowArms() {
@@ -644,7 +644,7 @@ public class EntityArmorStand extends EntityLivingBase {
             b0 = (byte) (b0 & -9);
         }
 
-        this.dataWatcher.updateObject(10, Byte.valueOf(b0));
+        this.dataWatcher.updateObject(10, b0);
     }
 
     public boolean hasNoBasePlate() {
@@ -660,7 +660,7 @@ public class EntityArmorStand extends EntityLivingBase {
             b0 = (byte) (b0 & -17);
         }
 
-        this.dataWatcher.updateObject(10, Byte.valueOf(b0));
+        this.dataWatcher.updateObject(10, b0);
     }
 
     public boolean hasMarker() {
