+++ src/main/java/net/optifine/util/ArrayUtils.java
@@ -136,7 +136,7 @@ public class ArrayUtils {
                     stringbuffer.append(separator);
                 }
 
-                stringbuffer.append(String.format(format, new Object[]{Float.valueOf(f)}));
+                stringbuffer.append(String.format(format, new Object[]{f}));
             }
 
             return stringbuffer.toString();
